ML model/requirements.txt
fastapi
uvicorn
pillow
tensorflow
numpy
